accounting: fix edge case in aged balance report detailed mode
If a transaction has the same source section and destination section, there was a risk of showing the wrong side of the transaction. Test that both section and nodes are the same instead. There is still a risk, of showing something wrong if source and destination are also same, but such transaction would be non sense anyway.