ERP5Form: relation fields must reindex when updating documents
Before, reindex was not called when you changed only a [Multi] relation field. So
it was possible to have SQL data not up to date after a change by user.

Though, most of the time this was not an issue. Usually relation fields
calls Base_edit. And even if there is no change and no reindex done by edit itself, since we
often have edit_workflow associated to documents, we usually have a reindexing.

But since we should keep the usage of edit_workflow optional, we have to call proper
setter to make sure to have updated SQL data.