core: avoid long urls with Base_jumpToRelatedObject
when script was called with "related=False", and when there was too many objects, then
a very long list of uid was given in url, and this could be truncated by web browser or web servers.
This was generating errors.

So instead, store the list of uids in the selection.