erp5_wendelin: Fix thinko in DataStream_copyCSVToDataArray() ZBigArray init
When we conditionally create new BigArray for appending data, we should create
it as empty, because in DataStream_copyCSVToDataArray() creation is done lazily
only when destination array is not yet initialized and we anyway append data to
the array in the following code block.

Creating BigArray with initial shape of appending part will result in
destination array being longer than neccessary by first-appended-chunk length
with this-way-introduced extra header reading as all zeros.

Fix it.

/cc @Tyagov