DataArray: Add a note a lot of code is currently duplicated from ERP5's BigFile
a9596bc4 (DataArray: get array slice by HTTP Range Request) added HTTP
range support for DataArray by way of copying code from ERP5's BigFile
implementation.

It is better to not copy the code and just use the base class and hook
into it in a proper way (I think it is possible to do even without
mixin).

Add a proper FIXME/TODO so we do not forget to fix it one day.

/reviewed-by @Tyagov
/cc @klaus