Make BusinessProcess temporary SM a child of its AR rather than Trade Model Path.
getTradePhaseMovementList() returns temporary SM with appropriate arrow
and dates but these temporary SM have their corresponding Trade Model
Path as a parent which means that Simulation Movement API is partially
unuseable (such as getting parent Applied Rule...).

Moreover, it seems strange for a Simulation Movement to not be a child
of an Applied Rule.

Therefore, create the temporary SM as a child of its corresponding AR.