[erp5_web_renderjs_ui] Router gadget should handle application redirection
The router gadget handles all URL interaction (generation, parsing). It is also responsible to handle redirection (as redirecting for hash and HTML5 pushState is different).

Ensure that redirect "block" the application until cancellation of the current page rendering.
This will prevent unexpected result for function calling redirect (in case of 401 for example)