BusinessTemplate: make sure we could modify broken objects before installation
Like we already do for Documents, Tests and Extension, we are soon
going to export as two files various types of document. It's usually
one file for metadata, and one file for the content (code, text).

Sometimes, we might want such kind of two file export for some  bt
having portal type and instance of such portal type in the same time.
In such case, we have some intermediate steps where we have broken
objects, and we need to change theses broken objets before portal
type is fully initialized and installed.