simulation: stop generating simulation movement ids longer at every split
When a delivery was splitted, then the new delivery was also splitted,
and so on, we were generiting ids like "1_split_0_split_0_split_0_split_0...".
It ended having simulation having a path up to 240 characters on some projects
due to this (near 255 catalog limit).

For keeping some information on the split, only keep "_split_" one time in the
id. If we really need to know exactly from which simulation movement a new
splitted simulation movement is coming from, we should use something else than
id.