Jupyter: automatic rendering of errors and objects as HTML
Now the ERP5 Jupyter kernel automatically renders errors that happens in the
user-side code. Errors are captured during the AST tree creation (to be able to
detect syntax errors) and at execution time.

Rendering of last-returning objects was implemented using a custom display hook.
The display hook is called everytime an object is executed in interactive mode.
To render a object of a given class a processor needs to be created and
associated with that class. When the class has no processor, the default one is
used to render the object's string representation. Many processors were created
for some special classes: Image class from the Image Module, Figure class from
matplotlib and all the IPython's display objects classes (Audio, IFrame,
YouTubeVideo, VimeoVideo, ScribdDocument, FileLink, and FileLinks).

An external method named Base_renderAsHtml was created to render objects that
are not the last-returning of their cells. It uses those already mentioned
processors for the rendering.

The user can manually add processors as he likes using the `_processor_list`
object that is automatically injected into the code context