erp5_upgrader: the constraint on Category Tool moves on Template Tool
Category Tool is very special for an ERP5, because it is part of the dynamic
generation of objects' properties. Setting up a property sheet on it will break
ERP5.

As the renaming feature is related to the upgrade, we will attach the constraint
on the Template Tool, as a way to group all the upgrade features at the same place