accounting: consider mirror node as an extra grouping parameter
Even if this is incorrect to allow grouping lines where mirror account does not match, we want to make it possible for sites where mirror account is not used consistently to group without having to migrate the data. So we make this possible by leaving them a chance to customize AccountingTransactionLine_getGroupingExtraParameterList to ignore mirror account.