catalog/stock: make stock table supporting movements indexed as several movements
The business template erp5_calendar had a customized version of z_catalog_stock_list
to have possibility to index a movement like it was several movements through method
asMovementList. This is particularly needed for group calendar assignments which needs
to insert as many movements as there is periods in a given time range.

So instead of keeping it as a duplicated z_catalog_stock_list, include in the generic
z_catalog_stock_list the generic concept of one movement indexed as several movements.

This should change nothing for existing usage of stock table, since asMovementList returns (self,)
for about all movements