erp5_catalog: New ERP5CatalogTool based on BaseTool from ERP5Type
Also, update in BusinessTemplate installation with changes in portal_catalog

Better to create new module for new Catalog Tool, as we don't
want a running instance to fail just because it is updated by
the new catalog code. This will maintain consistency in case
user doen't want to migrate to new catalog/catalog_tool

Most of the code for ERP5CatalogTool is copied from CatalogTool.