erp5_catalog: Conversion script to convert ZSQL Method and PythonScript to ERP5 objects
Also, Update BusinessTemplate installation with updated filter_dict
This removes the need to copy-patch or if-else on meta_type of catalog.
Use dynamic migration while installing the catalog method objects for
bt5.

Note : here we use conversion method for python script but dynamic migration for
ZSQL Methods. This is necessary because after the conversion to erp5 object, we need
to compile the body to generate code object for python script.
By just changing the class, this was not possible.

erp5_catalog: Simplify bt5 installation