erp5_catalog: Migration script to move catalog to new ERP5Catalog object.
What we expect during migration:
1. Copy objects as well as properties from SQLCatalog to ERP5Catalog

We need objects as well as properties to be setup because objects inside Catalogs
are basically zsql_methods which is what is needed when we use the properties.

For example: making call to `new_erp5_catalog.sql_clear_catalog` would return
a tuple of zsql methods which are basically catalog objects.
Right now, we are overwriting everything from SQLCatalog to the new ERP5Catalog without
checking for any conflicts(which is why there might be some overwrite).
This is why its recommended to call the migration function only during ERP5Site setup.

NOTE : For debugging purposes, we will be using in while installing and uninstallign BT5.
TODO : FIx overwriting so to make the migration function a bit more generic and use it anytime-anywhere.