ERP5 Subcategory: Remove '_'(underscore) from 'multiple_selection' subcategory in elmentary_type category.
Subcategory 'multiple_selection' has been renamed to 'multiple selection' to maintain consistency.
Earlier, whenever a new StandardCategory object was created with property_type 'multiple selection',
one needed to change the elementary type explicilty to use it to generate property setters and getters.
The example for this can be seen from this commit:
https://lab.nexedi.com/nexedi/erp5/commit/fd7387538a120307317c6c81cbe2399a866cdbb6
After this change, in the view, property type field couldn't recognize the property and displays
(???multiple selection).

Also, everywhere in erp5, list_types do mention 'multiple selction' and not 'multiple_selection'.
So, its better to maintain consistency in naming the subcategory.