erp5_catalog: Add select_variable property for Standard Property object in ERP5
This change solves the error which one gets while trying to access properties
Form for an ERP5 object having one or more of its property of type 'selection'
or 'multiple selection'.
For example: If you try to access property form for any portal_type objects,
you'll get an error : 'select_variable' is not defined. This is because of the
absence of the property select variable for the property of type multiple selection
which in case of portal_type object is property sheet list.

This is helpful for ERP5 views/forms which displays selection or multiple selection
type object.