Inventory API: drop default group by in getMovementHistoryList
Over the years, we included almost all possible default group by
criterions for getMovementHistoryList, so that in essence, we do not
group. A better solution is to just not group by by default, this
achieve the same and we do not have to filter duplicate in SQL (this
filter operation is expensive with large data sets).