erp5_catalog: [Performance] Concatenate all the queries into one while cataloging objects
Earlier, we used to make separate calls to valid catalog methods everytime
to catalog objects. It might be better to join all the query strings into
one string and make a single call.

Result on performance: Not much, cause its not the execution of catalog methods
which was eating time, but the rendering(method(src__=1, **kw)). Maybe check
the effect of this on high number of clusters as suggested by \@seb