erp5_catalog: Add typeBaseMethod(s) for PythonScript and SQLMethod portal types
The methods <portal_type>_getRedirectParameterDictAfterAdd are required for these
portal_types, cause their __call__ methods are overridden, so after addition
of new objects for them, it was taking to the url concerning wherever __call__
methods directed them.

Now, we change them to redirect to 'abosulte_url+'/view'' which is basically
the homepage for these catalog methods.