MailHost: Set SMTP socket timeout to 16s (#20161019-4A3BD2).
With old versions of Zope, it was possible to define the SMTP connection
timeout but this is not possible anymore even though MailHost, in turn
using zope.sendmail, uses Python smtplib library.

This meant that once connect() was called, it may have taken around 2
minutes or more before timeout (and this is OS-specific as well)
blocking one Zope thread.