'Standard Property' must be in 'property' Portal Type Group since 687fedd (#20161014-1C29326).
Otherwise, reset is not triggered when changing a 'Standard Property'
ZODB Property Sheet (through dynamic_class_generation_interaction_workflow
ChangeProperty_resetDynamicClasses Interaction). Also, add Unit Tests to
check that this Interaction is properly called.

'Base Category' was also removed in 687fedd but this does not seem to be
an issue as 'Category Property' and 'Dynamic Category Property' only use
the ID, and thus this is handled by PropertySheet_resetDynamicClasses
Interaction (_delObject/_setObject will be called).