erp5.util.testnode: make --frontend_url/node_quantity optional for runUnitTest
The recently added --frontend_url parameter breaks all existing runUnitTest
scripts that don't accept it, and because this option is irrelevant for many
of them, we don't want to force everyone to update their test suites.

Similarly, parallelizing tests is not worth the effort for some test suites,
and currently, their runUnitTest scripts contain ugly lines like:

    parser.add_argument('--node_quantity', help='ignored', type=int)

These 2 parameters are now passed to runUnitTest only if it supports them,
like it is already the case for --firefox_bin and --xvfb_bin.