Accounting: fix a wrong constraint when checking accounting periods
When we have a transaction with both sections from the same group, we
only check accounting periods for the sides which actually define
accounting. When there was a source or destination organiation defined
at the delivery level it was acquired on the lines and in that check we
wrongly considered this to be defining accounting.